Added statistic protection against preemption
Added a statistics-based protection against preemption. First,
latency of BT clock + SO clock reading is measured several times,
and a good quality average is calculated. Then, if latency of
clock reading is too high (compared with average), a new read
is taken because it is possible that there was a preemption
between two system calls.