Fixed device id calculation, added dependency to PyBlueZ CVS
Device ID determination for a given connection uses hci_get_route(),
which has been added in PyBlueZ CVS only, so HDPy now depends on
PyBlueZ >= 0.19 (not yet released) to work.

Some other low-level constants that were absent in PyBlueZ 0.18
and were hardcoded in HDPy could finally be removed, in the wake
of this dependence.