Added optionality to MDL reconnection feature
This commit makes MDL reconnection support optional. When reconnection
is disabled, MDLs that are closed or aborted are immediately deleted,
and reconnections are reject with 'not supported' response.

Also, the MCL class deletion method has been refactored; it notifies
the observer by itself and handles the "wildcard" MDL code, which
leads to a nice simplification in state machine code.